Gameplay tweaks / refacto
