+++ CMakeLists.txt
@@ -21,6 +21,8 @@ add_executable(Djipi-App
     "${CMAKE_CURRENT_SOURCE_DIR}/src/App.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Player.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Fireball.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Enemy.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/EnemyFactory.cpp"
 
     "${CMAKE_CURRENT_SOURCE_DIR}/Djipi/core/renderer/Renderer.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/Djipi/core/resource/ResourceManager.cpp"
